+++ modules/nlu/src/views/full/common/AutotrainToggle.tsx
@@ -2,7 +2,7 @@ import { Button } from '@blueprintjs/core'
 import { NLUApi } from 'api'
 import React, { FC, useEffect, useState } from 'react'
 
-import style from '../style.scss'
+import style from './style.scss'
 
 const AutotrainToggle: FC<{ api: NLUApi }> = ({ api }) => {
   const [autotrain, setAutotrain] = useState(false)
