+++ modules/nlu/src/views/full/common/TrainNow.tsx
