+++ modules/nlu/src/views/full/common/style.scss
@@ -1,0 +1,8 @@
+.trainingControl {
+  padding: 5px;
+  margin-bottom: 10px;
+}
+
+.autotrainToggle {
+  margin-right: 6px;
+}
