+++ modules/nlu/src/views/full/common/style.scss.d.ts
@@ -1,0 +1,8 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'autotrainToggle': string;
+  'trainingControl': string;
+}
+declare var cssExports: CssExports;
+export = cssExports;
