+++ modules/nlu/src/views/full/index.tsx
@@ -6,11 +6,11 @@ import React, { FC, useEffect, useState } from 'react'
 
 import { makeApi } from '../api'
 
+import TrainingControl from './common/TrainingControl'
 import EntityEditor from './entities/EntityEditor'
 import { EntitySidePanelSection } from './entities/SidePanelSection'
 import { IntentEditor } from './intents/FullEditor'
 import { IntentSidePanelSection } from './intents/SidePanelSection'
-import TrainingControl from './intents/TrainingControl'
 import style from './style.scss'
 
 export interface NluItem {
