+++ modules/nlu/src/views/full/style.scss
@@ -188,15 +188,6 @@ $status-bar-height: 24px;
   padding: 10px;
 }
 
-.trainingControl {
-  padding: 5px;
-  margin-bottom: 10px;
-}
-
-.autotrainToggle {
-  margin-right: 6px;
-}
-
 .trainingControlContainer {
   display: flex;
   flex-direction: row-reverse;
